+++ packages/web3-modal/src/components/modals/ConfigTypeModal/index.tsx
@@ -105,4 +105,5 @@ function NetworkModalContent() {
   )
 }
 
+// Exported as default for lazy import inside /components/modals/index.tsx
 export default memo(NetworkModalContent)