+++ packages/web3-modal/src/components/modals/common/PoweredByLabel.tsx
@@ -1,10 +1,21 @@
 import { Row } from '@past3lle/components'
-import { setBestTextColour } from '@past3lle/theme'
 import { MakeOptional } from '@past3lle/types'
 import React from 'react'
 import styled from 'styled-components'
 
 import { ContainerThemeTypes, ModalText } from './styled'
+
+const AttributionText = styled(ModalText).attrs((props: any) => ({
+  color: props.theme.modals?.base?.attribution?.color
+}))`
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  gap: 5px;
+  > img {
+    width: 40px;
+  }
+`
 
 const PoweredByLabelContainer = styled(Row).attrs({
   justifyContent: 'center',
@@ -13,19 +24,6 @@ const PoweredByLabelContainer = styled(Row).attrs({
   padding: '3px'
 })<ContainerThemeTypes>`
   z-index: 500;
-  > ${ModalText} {
-    color: ${({ theme: { modals }, modal }) =>
-      modals?.[modal]?.background?.main
-        ? setBestTextColour(modals?.base?.background?.main as string, 7)
-        : modals?.[modal]?.font?.color};
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    gap: 5px;
-    > img {
-      width: 40px;
-    }
-  }
 
   a {
     color: inherit;
@@ -42,14 +40,14 @@ const PoweredByLabelContainer = styled(Row).attrs({
 export function PoweredByLabel({ modal = 'base', node = 'main' }: MakeOptional<ContainerThemeTypes, 'modal' | 'node'>) {
   return (
     <PoweredByLabelContainer modal={modal} node={node} title="POWERED BY PAST3LLE LABS">
-      <ModalText modal="base" node="subHeader" fontSize={10}>
+      <AttributionText modal="base" node="subHeader" fontSize={10}>
         <a href="https://pastellelabs.com?=referrer=pstl-web3-modal" target="_blank" referrerPolicy="no-referrer">
           POWERED BY{' '}
-          <ModalText modal="base" node="strong" display="inline-block">
+          <AttributionText modal="base" node="strong" display="inline-block">
             PAST3LLE LABS
-          </ModalText>
+          </AttributionText>
         </a>
-      </ModalText>
+      </AttributionText>
     </PoweredByLabelContainer>
   )
 }