+++ apps/pastellelabs-landing-ui/CHANGELOG.md
@@ -3,6 +3,54 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.1-alpha.5](https://github.com/PAST3LLE/past3lle-monorepo/compare/@past3lle/pastellelabs-landing-ui@0.2.1-alpha.4...@past3lle/pastellelabs-landing-ui@0.2.1-alpha.5) (2024-03-01)
+
+**Note:** Version bump only for package @past3lle/pastellelabs-landing-ui
+
+
+
+
+
+## [0.2.1-alpha.4](https://github.com/PAST3LLE/past3lle-monorepo/compare/@past3lle/pastellelabs-landing-ui@0.2.1-alpha.3...@past3lle/pastellelabs-landing-ui@0.2.1-alpha.4) (2024-02-29)
+
+**Note:** Version bump only for package @past3lle/pastellelabs-landing-ui
+
+
+
+
+
+## [0.2.1-alpha.3](https://github.com/PAST3LLE/past3lle-monorepo/compare/@past3lle/pastellelabs-landing-ui@0.2.1-alpha.2...@past3lle/pastellelabs-landing-ui@0.2.1-alpha.3) (2024-02-28)
+
+**Note:** Version bump only for package @past3lle/pastellelabs-landing-ui
+
+
+
+
+
+## [0.2.1-alpha.2](https://github.com/PAST3LLE/past3lle-monorepo/compare/@past3lle/pastellelabs-landing-ui@0.2.1-alpha.1...@past3lle/pastellelabs-landing-ui@0.2.1-alpha.2) (2024-02-28)
+
+**Note:** Version bump only for package @past3lle/pastellelabs-landing-ui
+
+
+
+
+
+## [0.2.1-alpha.1](https://github.com/PAST3LLE/past3lle-monorepo/compare/@past3lle/pastellelabs-landing-ui@0.2.1-alpha.0...@past3lle/pastellelabs-landing-ui@0.2.1-alpha.1) (2024-01-31)
+
+**Note:** Version bump only for package @past3lle/pastellelabs-landing-ui
+
+
+
+
+
+## [0.2.1-alpha.0](https://github.com/PAST3LLE/past3lle-monorepo/compare/@past3lle/pastellelabs-landing-ui@0.2.0...@past3lle/pastellelabs-landing-ui@0.2.1-alpha.0) (2024-01-30)
+
+**Note:** Version bump only for package @past3lle/pastellelabs-landing-ui
+
+
+
+
+
 # [0.2.0](https://github.com/PAST3LLE/past3lle-monorepo/compare/@past3lle/pastellelabs-landing-ui@0.1.14...@past3lle/pastellelabs-landing-ui@0.2.0) (2024-01-26)