+++ apps/skillforge-ui/CHANGELOG.md
@@ -3,6 +3,57 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.1.1-alpha.5](https://github.com/PAST3LLE/past3lle-monorepo/compare/@past3lle/skillforge-ui@1.1.1-alpha.4...@past3lle/skillforge-ui@1.1.1-alpha.5) (2024-03-01)
+
+**Note:** Version bump only for package @past3lle/skillforge-ui
+
+
+
+
+
+## [1.1.1-alpha.4](https://github.com/PAST3LLE/past3lle-monorepo/compare/@past3lle/skillforge-ui@1.1.1-alpha.3...@past3lle/skillforge-ui@1.1.1-alpha.4) (2024-02-29)
+
+**Note:** Version bump only for package @past3lle/skillforge-ui
+
+
+
+
+
+## [1.1.1-alpha.3](https://github.com/PAST3LLE/past3lle-monorepo/compare/@past3lle/skillforge-ui@1.1.1-alpha.2...@past3lle/skillforge-ui@1.1.1-alpha.3) (2024-02-28)
+
+**Note:** Version bump only for package @past3lle/skillforge-ui
+
+
+
+
+
+## [1.1.1-alpha.2](https://github.com/PAST3LLE/past3lle-monorepo/compare/@past3lle/skillforge-ui@1.1.1-alpha.1...@past3lle/skillforge-ui@1.1.1-alpha.2) (2024-02-28)
+
+**Note:** Version bump only for package @past3lle/skillforge-ui
+
+
+
+
+
+## [1.1.1-alpha.1](https://github.com/PAST3LLE/past3lle-monorepo/compare/@past3lle/skillforge-ui@1.1.1-alpha.0...@past3lle/skillforge-ui@1.1.1-alpha.1) (2024-01-31)
+
+
+### Bug Fixes
+
+* **skillforge-ui,forge-web3,skillforge-widget,wagmi-connectors,web3-modal:** bump & sync Wagmi/Viem/react-query package versions ([#66](https://github.com/PAST3LLE/past3lle-monorepo/issues/66)) ([141edbd](https://github.com/PAST3LLE/past3lle-monorepo/commit/141edbde34b5021e05c58569e545dc4a0a28768b))
+
+
+
+
+
+## [1.1.1-alpha.0](https://github.com/PAST3LLE/past3lle-monorepo/compare/@past3lle/skillforge-ui@1.1.0...@past3lle/skillforge-ui@1.1.1-alpha.0) (2024-01-30)
+
+**Note:** Version bump only for package @past3lle/skillforge-ui
+
+
+
+
+
 # [1.1.0](https://github.com/PAST3LLE/past3lle-monorepo/compare/@past3lle/skillforge-ui@1.0.0-alpha.3...@past3lle/skillforge-ui@1.1.0) (2024-01-26)